wanda-voice/requirements.txt
# Core dependencies
faster-whisper>=1.1.0
edge-tts>=6.1.0  # Microsoft Neural TTS (Siri-Qualität)
piper-tts>=1.2.0  # Offline Fallback
sounddevice==0.5.5
scipy>=1.17.0
numpy>=2.0
evdev==1.9.2
pyyaml>=6.0

# JARVIS Phase 2: SOTA VAD + Session
torch>=2.0
pexpect>=4.9

# Phase 3: HTTP for Ollama
requests>=2.31

# GTK3 for visual orb (via system: sudo apt install python3-gi)
PyGObject>=3.42

# Optional: Silero VAD (included with torch)
# silero-vad>=5.1

# Optional: Wake Word
# openwakeword>=0.6.0

# Mobile Access (Telegram Bot)
python-telegram-bot>=21.0

# Note: ffmpeg required for voice message conversion
# Install: sudo apt install ffmpeg
